requirements.txt
torch>=2.2.0
torchvision>=0.13.0
diffusers>=0.20.0,<0.25.0
transformers>=4.30.0,<4.35.0
accelerate>=0.20.0,<0.24.0
gradio>=3.40.0,<3.51.0
pillow>=9.0.0
numpy>=1.26.0
tqdm>=4.65.0
huggingface-hub>=0.16.4,<0.19.0